+++ pkg/rpc/utils.go
@@ -1,7 +1,6 @@
 package rpc
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"time"
@@ -221,50 +220,4 @@ func validatePage(pagePtr *int, perPage, totalCount int) (int, error) {
 	}
 
 	return page, nil
-}
-
-func (r *RPCServer) normalizeHeight(height *int64) uint64 {
-	var heightValue uint64
-	if height == nil {
-		var err error
-		// TODO: Decide how to handle context here. Using background for now.
-		heightValue, err = r.adapter.Store.Height(context.Background())
-		if err != nil {
-			// TODO: Consider logging or returning error
-			return 0
-		}
-	} else if *height < 0 {
-		// Handle negative heights if they have special meaning (e.g., -1 for latest)
-		// Currently, just treat them as 0 or latest, adjust as needed.
-		// For now, let's assume negative height means latest valid height.
-		var err error
-		heightValue, err = r.adapter.Store.Height(context.Background())
-		if err != nil {
-			return 0
-		}
-	} else {
-		heightValue = uint64(*height)
-	}
-
-	return heightValue
-}
-
-func (r *RPCServer) getBlockMeta(ctx context.Context, n uint64) *cmtypes.BlockMeta {
-	header, data, err := r.adapter.Store.GetBlockData(ctx, n)
-	if err != nil {
-		// TODO: Log error? Return error?
-		return nil
-	}
-	// Handle case where GetBlockData might return nil header/data without error?
-	if header == nil || data == nil {
-		// TODO: Log this situation?
-		return nil
-	}
-	bmeta, err := ToABCIBlockMeta(header, data)
-	if err != nil {
-		// TODO: Log error? Return error?
-		return nil
-	}
-
-	return bmeta
 }
